+++ testing/src/main/kotlin/org/shedlang/compiler/tests/builders.kt
@@ -107,7 +107,7 @@ fun literalBool(
     value: Boolean = false,
     source: Source = anySource()
 ) = BooleanLiteralNode(value, source = source)
-fun literalInt(value: Int = 0) = IntegerLiteralNode(value, anySource())
+fun literalInt(value: Int = 0) = IntegerLiteralNode(value.toBigInteger(), anySource())
 fun literalString(value: String = "") = StringLiteralNode(value, anySource())
 fun literalChar(value: Char = '!') = CharacterLiteralNode(value.toInt(), anySource())
 fun symbolName(name: String) = SymbolNode(name, anySource())
